// src/app/model/jung/productVO.ts
export interface ProductVO {
    PT_CODE?: string;
    USER_ID?: string;
    PRODUCT_CD?: string;
    PRODUCT_TYPE?: string;
    PRODUCT_STATE?: string;
    PRODUCT_THS?: string;
    AP_DATE?: string;
    GD_DATE?: string;
    ST_COIN?: string;
    CRT_USD?: string;
    AUTO_SHIP?: string;
    REG_DATE?: string;
    REG_USER?: string;
    UP_DATE?: string;
    UP_USER?: string;
    USE_FLAG?: string;
    END_DATE?: string;
}